requirements.txt
Django==1.11.17
django-filter==2.0.0
djangorestframework==3.8.2
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
pandas==0.23.4
xlrd==1.1.0
mathjs==5.1.2
django-allauth==0.38.0
django-mailer==1.2.5
